min time climb example updates
