+++ dymos/examples/min_time_climb/doc/test_doc_min_time_climb.py
@@ -2,32 +2,48 @@ from __future__ import print_function, absolute_import, division
 
 import unittest
 
+import matplotlib
+matplotlib.use('Agg')
+import matplotlib.pyplot as plt
+plt.style.use('ggplot')
+
 
 class TestMinTimeClimbForDocs(unittest.TestCase):
 
     def test_min_time_climb_for_docs_gauss_lobatto(self):
-        import matplotlib
-        matplotlib.use('Agg')
         import matplotlib.pyplot as plt
 
         from openmdao.api import Problem, Group, pyOptSparseDriver, DirectSolver
         from openmdao.utils.assert_utils import assert_rel_error
 
-        from dymos import Phase, GaussLobatto
+        import dymos as dm
         from dymos.examples.min_time_climb.min_time_climb_ode import MinTimeClimbODE
+        from dymos.examples.plotting import plot_results
 
+        #
+        # Instantiate the problem and configure the optimization driver
+        #
         p = Problem(model=Group())
 
         p.driver = pyOptSparseDriver()
         p.driver.options['optimizer'] = 'SLSQP'
         p.driver.options['dynamic_simul_derivs'] = True
 
-        phase = Phase(ode_class=MinTimeClimbODE,
-                      transcription=GaussLobatto(num_segments=12,
-                                                 compressed=True))
+        #
+        # Instantiate the trajectory and phase
+        #
+        traj = dm.Trajectory()
 
-        p.model.add_subsystem('phase0', phase)
+        phase = dm.Phase(ode_class=MinTimeClimbODE,
+                         transcription=dm.GaussLobatto(num_segments=20, compressed=True))
 
+        traj.add_phase('phase0', phase)
+
+        p.model.add_subsystem('traj', traj)
+
+        #
+        # Set the options on the optimization variables
+        #
         phase.set_time_options(fix_initial=True, duration_bounds=(50, 400),
                                duration_ref=100.0)
 
@@ -54,62 +70,57 @@ class TestMinTimeClimbForDocs(unittest.TestCase):
         phase.add_design_parameter('Isp', val=1600.0, units='s', opt=False)
         phase.add_design_parameter('throttle', val=1.0, opt=False)
 
+        #
+        # Setup the boundary and path constraints
+        #
         phase.add_boundary_constraint('h', loc='final', equals=20000, scaler=1.0E-3, units='m')
         phase.add_boundary_constraint('aero.mach', loc='final', equals=1.0, shape=(1,))
         phase.add_boundary_constraint('gam', loc='final', equals=0.0, units='rad')
 
         phase.add_path_constraint(name='h', lower=100.0, upper=20000, ref=20000)
         phase.add_path_constraint(name='aero.mach', lower=0.1, upper=1.8, shape=(1,))
-        phase.add_path_constraint(name='alpha', lower=-8, upper=8)
 
         # Minimize time at the end of the phase
         phase.add_objective('time', loc='final', ref=1.0)
 
         p.model.linear_solver = DirectSolver()
 
+        #
+        # Setup the problem and set the initial guess
+        #
         p.setup(check=True)
 
-        p['phase0.t_initial'] = 0.0
-        p['phase0.t_duration'] = 500
+        p['traj.phase0.t_initial'] = 0.0
+        p['traj.phase0.t_duration'] = 500
 
-        p['phase0.states:r'] = phase.interpolate(ys=[0.0, 50000.0], nodes='state_input')
-        p['phase0.states:h'] = phase.interpolate(ys=[100.0, 20000.0], nodes='state_input')
-        p['phase0.states:v'] = phase.interpolate(ys=[135.964, 283.159], nodes='state_input')
-        p['phase0.states:gam'] = phase.interpolate(ys=[0.0, 0.0], nodes='state_input')
-        p['phase0.states:m'] = phase.interpolate(ys=[19030.468, 10000.], nodes='state_input')
-        p['phase0.controls:alpha'] = phase.interpolate(ys=[0.0, 0.0], nodes='control_input')
+        p['traj.phase0.states:r'] = phase.interpolate(ys=[0.0, 50000.0], nodes='state_input')
+        p['traj.phase0.states:h'] = phase.interpolate(ys=[100.0, 20000.0], nodes='state_input')
+        p['traj.phase0.states:v'] = phase.interpolate(ys=[135.964, 283.159], nodes='state_input')
+        p['traj.phase0.states:gam'] = phase.interpolate(ys=[0.0, 0.0], nodes='state_input')
+        p['traj.phase0.states:m'] = phase.interpolate(ys=[19030.468, 10000.], nodes='state_input')
+        p['traj.phase0.controls:alpha'] = phase.interpolate(ys=[0.0, 0.0], nodes='control_input')
 
+        #
         # Solve for the optimal trajectory
+        #
         p.run_driver()
 
+        #
         # Test the results
-        assert_rel_error(self, p.get_val('phase0.t_duration'), 321.0, tolerance=1.0E-1)
+        #
+        assert_rel_error(self, p.get_val('traj.phase0.t_duration'), 321.0, tolerance=1.0E-1)
 
-        # Verify that ODE output mach is added to the timeseries
-        assert_rel_error(self, p.get_val('phase0.timeseries.mach')[-1], 1.0, tolerance=1.0E-2)
+        #
+        # Get the explicitly simulated solution and plot the results
+        #
+        exp_out = traj.simulate()
 
-        exp_out = phase.simulate()
-
-        fig, axes = plt.subplots(2, 1, sharex=True)
-
-        t_sol = p.get_val('phase0.timeseries.time')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-
-        h_sol = p.get_val('phase0.timeseries.states:h')
-        h_exp = exp_out.get_val('phase0.timeseries.states:h')
-
-        alpha_sol = p.get_val('phase0.timeseries.controls:alpha', units='deg')
-        alpha_exp = exp_out.get_val('phase0.timeseries.controls:alpha', units='deg')
-
-        axes[0].plot(t_sol, h_sol, 'ro')
-        axes[0].plot(t_exp, h_exp, 'b-')
-        axes[0].set_xlabel('time (s)')
-        axes[0].set_ylabel('altitude (m)')
-
-        axes[1].plot(t_sol, alpha_sol, 'ro')
-        axes[1].plot(t_exp, alpha_exp, 'b-')
-        axes[1].set_xlabel('time (s)')
-        axes[1].set_ylabel('alpha (deg)')
+        plot_results([('traj.phase0.timeseries.time', 'traj.phase0.timeseries.states:h',
+                       'time (s)', 'altitude (m)'),
+                      ('traj.phase0.timeseries.time', 'traj.phase0.timeseries.controls:alpha',
+                       'time (s)', 'alpha (deg)')],
+                     title='Supersonic Minimum Time-to-Climb Solution',
+                     p_sol=p, p_sim=exp_out)
 
         plt.show()
