--- dymos/examples/min_time_climb/ex_min_time_climb.py
@@ -1,138 +1,0 @@
-from __future__ import print_function, division, absolute_import
-
-import numpy as np
-
-from openmdao.api import Problem, Group, pyOptSparseDriver, DirectSolver
-
-from dymos import Phase, GaussLobatto, Radau, RungeKutta
-
-from dymos.examples.min_time_climb.min_time_climb_ode import MinTimeClimbODE
-
-import matplotlib
-matplotlib.use('Agg')
-
-SHOW_PLOTS = True
-
-
-def min_time_climb(optimizer='SLSQP', num_seg=3, transcription='gauss-lobatto',
-                   transcription_order=3, force_alloc_complex=False):
-
-    p = Problem(model=Group())
-
-    p.driver = pyOptSparseDriver()
-    p.driver.options['optimizer'] = optimizer
-    p.driver.options['dynamic_simul_derivs'] = True
-
-    if optimizer == 'SNOPT':
-        p.driver.opt_settings['Major iterations limit'] = 1000
-        p.driver.opt_settings['iSumm'] = 6
-        p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-6
-        p.driver.opt_settings['Major optimality tolerance'] = 1.0E-6
-        p.driver.opt_settings['Function precision'] = 1.0E-12
-        p.driver.opt_settings['Linesearch tolerance'] = 0.1
-        p.driver.opt_settings['Major step limit'] = 0.5
-        # p.driver.opt_settings['Verify level'] = 3
-
-    t = {'gauss-lobatto': GaussLobatto(num_segments=num_seg, order=transcription_order),
-         'radau-ps': Radau(num_segments=num_seg, order=transcription_order),
-         'runge-kutta': RungeKutta(num_segments=num_seg)}
-
-    phase = Phase(ode_class=MinTimeClimbODE, transcription=t[transcription])
-
-    p.model.add_subsystem('phase0', phase)
-
-    phase.set_time_options(fix_initial=True, duration_bounds=(50, 400),
-                           duration_ref=100.0)
-
-    phase.set_state_options('r', fix_initial=True, lower=0, upper=1.0E6,
-                            ref=1.0E3, defect_ref=1.0E3, units='m')
-
-    phase.set_state_options('h', fix_initial=True, lower=0, upper=20000.0,
-                            ref=1.0E2, defect_ref=1.0E2, units='m')
-
-    phase.set_state_options('v', fix_initial=True, lower=10.0,
-                            ref=1.0E2, defect_ref=1.0E2, units='m/s')
-
-    phase.set_state_options('gam', fix_initial=True, lower=-1.5, upper=1.5,
-                            ref=1.0, defect_ref=1.0, units='rad')
-
-    phase.set_state_options('m', fix_initial=True, lower=10.0, upper=1.0E5,
-                            ref=1.0E3, defect_ref=1.0E3)
-
-    phase.add_control('alpha', units='deg', lower=-8.0, upper=8.0, scaler=1.0,
-                      rate_continuity=True, rate_continuity_scaler=100.0,
-                      rate2_continuity=False)
-
-    phase.add_design_parameter('S', val=49.2386, units='m**2', opt=False)
-    phase.add_design_parameter('Isp', val=1600.0, units='s', opt=False)
-    phase.add_design_parameter('throttle', val=1.0, opt=False)
-
-    phase.add_boundary_constraint('h', loc='final', equals=20000, scaler=1.0E-3, units='m')
-    phase.add_boundary_constraint('aero.mach', loc='final', equals=1.0)
-    phase.add_boundary_constraint('gam', loc='final', equals=0.0, units='rad')
-
-    phase.add_path_constraint(name='h', lower=100.0, upper=20000, ref=20000)
-    phase.add_path_constraint(name='aero.mach', lower=0.1, upper=1.8)
-    phase.add_path_constraint(name='alpha', lower=-8, upper=8)
-
-    # Minimize time at the end of the phase
-    phase.add_objective('time', loc='final', ref=1.0)
-
-    p.model.linear_solver = DirectSolver()
-
-    p.setup(check=True, force_alloc_complex=force_alloc_complex)
-
-    p['phase0.t_initial'] = 0.0
-    p['phase0.t_duration'] = 300.0
-
-    p['phase0.states:r'] = phase.interpolate(ys=[0.0, 111319.54], nodes='state_input')
-    p['phase0.states:h'] = phase.interpolate(ys=[100.0, 20000.0], nodes='state_input')
-    p['phase0.states:v'] = phase.interpolate(ys=[135.964, 283.159], nodes='state_input')
-    p['phase0.states:gam'] = phase.interpolate(ys=[0.0, 0.0], nodes='state_input')
-    p['phase0.states:m'] = phase.interpolate(ys=[19030.468, 16841.431], nodes='state_input')
-    p['phase0.controls:alpha'] = phase.interpolate(ys=[0.0, 0.0], nodes='control_input')
-
-    p.run_driver()
-
-    if SHOW_PLOTS:
-        exp_out = phase.simulate()
-
-        t_sol = p.get_val('phase0.timeseries.time')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-
-        h_sol = p.get_val('phase0.timeseries.states:h')
-        h_exp = exp_out.get_val('phase0.timeseries.states:h')
-
-        v_sol = p.get_val('phase0.timeseries.states:v')
-        v_exp = exp_out.get_val('phase0.timeseries.states:v')
-
-        alpha_sol = p.get_val('phase0.timeseries.controls:alpha')
-        alpha_exp = exp_out.get_val('phase0.timeseries.controls:alpha')
-
-        import matplotlib.pyplot as plt
-        plt.plot(t_sol, h_sol, 'ro')
-        plt.plot(t_exp, h_exp, 'b-')
-        plt.xlabel('time (s)')
-        plt.ylabel('altitude (m)')
-
-        plt.figure()
-        plt.plot(v_sol, h_sol, 'ro')
-        plt.plot(v_exp, h_exp, 'b-')
-        plt.xlabel('airspeed (m/s)')
-        plt.ylabel('altitude (m)')
-
-        plt.figure()
-        plt.plot(t_sol, alpha_sol, 'ro')
-        plt.plot(t_exp, alpha_exp, 'b-')
-        plt.xlabel('time (s)')
-        plt.ylabel('alpha (rad)')\
-
-        plt.show()
-
-    return p
-
-
-if __name__ == '__main__':
-    SHOW_PLOTS = True
-    p = min_time_climb(
-        optimizer='SLSQP', num_seg=12, transcription='gauss-lobatto', transcription_order=3)
